chore: rename resources to projects
