+++ lib/core/router/router.dart
@@ -6,7 +6,7 @@ import 'package:datn_mobile/core/router/router.gr.dart';
 class AppRouter extends RootStackRouter {
   @override
   late final List<AutoRoute> routes = [
-    AutoRoute(page: ResourcesRoute.page, path: '/resources', children: []),
+    AutoRoute(page: ProjectsRoute.page, path: '/resources', children: []),
     AutoRoute(page: SettingRoute.page, path: '/settings', initial: true),
   ];
 }
