+++ lib/features/projects/domain/entity/value_object/slide_background.dart
