+++ lib/features/projects/domain/repository/presentation_repository.dart
@@ -1,0 +1,7 @@
+import 'package:datn_mobile/features/projects/domain/entity/presentation.dart';
+import 'package:datn_mobile/features/projects/domain/entity/presentation_minimal.dart';
+
+abstract class PresentationRepository {
+  Future<List<PresentationMinimal>> fetchPresentations();
+  Future<void> addPresentation(Presentation presentation);
+}
