+++ lib/features/projects/enum/slide_element_type.dart
