+++ lib/features/projects/ui/widgets/common/thumbnail.dart
