--- lib/features/resources/domain/entity/presentation_minimal.dart
@@ -1,33 +1,0 @@
-import 'package:datn_mobile/features/resources/domain/entity/value_object/slide.dart';
-
-class PresentationMinimal {
-  final String? id;
-  final String? title;
-  final DateTime? createdAt;
-  final DateTime? updatedAt;
-  final Slide? thumbnail;
-
-  const PresentationMinimal({
-    this.id,
-    this.title,
-    this.createdAt,
-    this.updatedAt,
-    this.thumbnail,
-  });
-
-  PresentationMinimal copyWith({
-    String? id,
-    String? title,
-    DateTime? createdAt,
-    DateTime? updatedAt,
-    Slide? thumbnail,
-  }) {
-    return PresentationMinimal(
-      id: id ?? this.id,
-      title: title ?? this.title,
-      createdAt: createdAt ?? this.createdAt,
-      updatedAt: updatedAt ?? this.updatedAt,
-      thumbnail: thumbnail ?? this.thumbnail,
-    );
-  }
-}
