+++ lib/shared/widget/bottom_app_bar.dart
@@ -8,9 +8,7 @@ class CustomBottomAppBar extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return AutoTabsScaffold(
-      routes: const [
-        ResourcesRoute(),
-      ],
+      routes: const [ProjectsRoute()],
       bottomNavigationBuilder: (_, tabsRouter) {
         return BottomNavigationBar(
           currentIndex: tabsRouter.activeIndex,
